enabled the ability to create tasks with positions in lists

+++ backend/src/main/java/com/killeen/taskflow/components/task/model/Task.java
@@ -22,7 +22,7 @@ public class Task {
     private boolean       completed;
     private LocalDate     dueDate;
     private List<Subtask> subtasks;
-    private Integer       position;
+    private Long          position;
     private OffsetDateTime createdAt;
     private OffsetDateTime updatedAt;
 }
